done all

+++ src/main/java/com/example/tshirt_luxury_datn/controller/banHangOnlController.java
@@ -4,11 +4,14 @@ import com.example.tshirt_luxury_datn.entity.*;
 import com.example.tshirt_luxury_datn.repository.*;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Date;
 import java.util.List;
@@ -62,6 +65,7 @@ public class banHangOnlController {
             model.addAttribute("tongTienGioHang", gioHangChiTietRepo.tinhTongGia(idGioHang));
             model.addAttribute("banHang", gioHangChiTiets);
             model.addAttribute("giaTriGiam", session.getAttribute("giaTriGiam"));
+            model.addAttribute("TTGHNull",session.getAttribute("TTGHNull"));
 
             return "BanHang/ban-hang-onl";
         }
@@ -69,11 +73,26 @@ public class banHangOnlController {
 
     @GetMapping("/t-shirt-luxury/ban-hang-onl/getVoucher")
     public String getVoucher(@RequestParam("idVoucher") Integer idVC, HttpSession session) {
+        Integer idGioHang = (Integer) session.getAttribute("idGioHang");
         Integer giaTriGiam = voucherRepo.getGiaTriGiam(idVC);
-
-        session.setAttribute("giaTriGiam", giaTriGiam);
+        Float tongTien =  gioHangChiTietRepo.tinhTongGia(idGioHang);
+        Float giamGia = tongTien * giaTriGiam / 100;
         Voucher voucher = voucherRepo.getReferenceById(idVC);
-        session.setAttribute("voucher", voucher);
+        if(voucher.getGioiHan()<=giamGia){
+            session.setAttribute("giaTriGiam", voucher.getGioiHan());
+            voucher.setSoLuong(voucher.getSoLuong()-1);
+            voucherRepo.save(voucher);
+            session.setAttribute("voucher", voucher);
+        }else {
+            session.setAttribute("giaTriGiam", giamGia);
+            voucher.setSoLuong(voucher.getSoLuong()-1);
+            voucherRepo.save(voucher);
+            session.setAttribute("voucher", voucher);
+        }
+
+
+
+
 
         return "redirect:/t-shirt-luxury/ban-hang-onl";
     }
@@ -105,21 +124,55 @@ public class banHangOnlController {
         return "redirect:/t-shirt-luxury/ban-hang-onl";
     }
 
-    @PostMapping("/t-shirt-luxury/ban-hang-onl/createHD")
-    public String createHD(HttpSession session, ThongTinNhanHang thongTinNhanHang) {
-        thongTinDonHangRepo.save(thongTinNhanHang);
-        Voucher voucher = voucherRepo.getReferenceById(1);
-        HoaDon hoaDon = new HoaDon();
-        hoaDon.setNgaySua(new Date());
-        hoaDon.setNgayTao(new Date());
-        hoaDon.setTrangThai(0);
-        hoaDon.setLoaiHoaDon(1);
-        hoaDon.setVoucher(voucher);
-        hoaDon.setThongTinNhanHang(thongTinNhanHang);
-        hoaDonRepo.save(hoaDon);
+    boolean checkTTNH(ThongTinNhanHang thongTinNhanHang, RedirectAttributes redirectAttributes) {
+        if (thongTinNhanHang.getHoVaTen().isEmpty() || thongTinNhanHang.getHoVaTen() == null) {
+            String validHoTen = "Họ tên không được để trống";
+            redirectAttributes.addFlashAttribute("errorName", validHoTen);
+            return false;
+        } else {
+            redirectAttributes.addFlashAttribute("hoVaTen", thongTinNhanHang.getHoVaTen());
+        }
+        if (thongTinNhanHang.getSoDienThoai().isEmpty() || thongTinNhanHang.getSoDienThoai() == null) {
+            String validSDT = "Số điện thoại không được để trống";
+            redirectAttributes.addFlashAttribute("errorNumber", validSDT);
+            return false;
+        } else {
+            redirectAttributes.addFlashAttribute("soDienThoai", thongTinNhanHang.getSoDienThoai());
+        }
+        if (thongTinNhanHang.getDiaChiNhanHang().isEmpty() || thongTinNhanHang.getDiaChiNhanHang() == null) {
+            String valiAddess = "Địa chỉ thoại không được để trống";
+            redirectAttributes.addFlashAttribute("errorAddess", valiAddess);
+            return false;
+        } else {
+            redirectAttributes.addFlashAttribute("diaChiNhanHang", thongTinNhanHang.getDiaChiNhanHang());
+        }
 
-        session.setAttribute("hoaDonOnl", hoaDon);
-        session.setAttribute("idHoaDonOnline", hoaDon.getId());
+        return true;
+
+    }
+
+    @PostMapping("/t-shirt-luxury/ban-hang-onl/createHD")
+    public String createHD(HttpSession session, ThongTinNhanHang thongTinNhanHang,
+                           RedirectAttributes redirectAttributes
+    ) {
+
+        if (checkTTNH(thongTinNhanHang, redirectAttributes)) {
+            thongTinDonHangRepo.save(thongTinNhanHang);
+            Voucher voucher = voucherRepo.getReferenceById(1);
+            HoaDon hoaDon = new HoaDon();
+            hoaDon.setNgaySua(new Date());
+            hoaDon.setNgayTao(new Date());
+            hoaDon.setTrangThai(0);
+            hoaDon.setLoaiHoaDon(1);
+            hoaDon.setVoucher(voucher);
+            hoaDon.setThongTinNhanHang(thongTinNhanHang);
+            hoaDonRepo.save(hoaDon);
+
+            session.setAttribute("hoaDonOnl", hoaDon);
+            session.setAttribute("idHoaDonOnline", hoaDon.getId());
+
+        }
+
 
         return "redirect:/t-shirt-luxury/ban-hang-onl";
     }
@@ -132,42 +185,69 @@ public class banHangOnlController {
         // Chuyển đổi sang Float
         float tongTienParsed = Float.parseFloat(tongTien);
 
+
         HoaDonChiTiet hoaDonChiTiet = new HoaDonChiTiet();
         HoaDon hoaDon = (HoaDon) session.getAttribute("hoaDonOnl");
-        hoaDon.setMaHoaDon(generateMaHoaDon());
-        hoaDon.setVoucher((Voucher) session.getAttribute("voucher"));
-        hoaDon.setTongTien(tongTienParsed);
-        hoaDon.setTrangThai(2);
-        hoaDonRepo.save(hoaDon);
-        Integer idGioHang = (Integer) session.getAttribute("idGioHang");
+        if (hoaDon == null || hoaDon.getThongTinNhanHang() == null || hoaDon.getThongTinNhanHang().getId() == null) {
+            String TTGHNull = "Chưa có thông tin giao hàng";
+            session.setAttribute("TTGHNull", TTGHNull);
+            return "redirect:/t-shirt-luxury/ban-hang-onl";
+        }
+        else {
 
-        List<GioHangChiTiet> gioHangChiTiets = gioHangChiTietRepo.gioHangChiTietByID(idGioHang);
-        List<Integer> listIDSPCT = gioHangChiTietRepo.findIdSanPhamChiTietByIdGioHang(idGioHang);
+            hoaDon.setMaHoaDon(generateMaHoaDon());
+            hoaDon.setVoucher((Voucher) session.getAttribute("voucher"));
+            hoaDon.setTongTien(tongTienParsed);
+            hoaDon.setTrangThai(2);
+            hoaDonRepo.save(hoaDon);
+            Integer idGioHang = (Integer) session.getAttribute("idGioHang");
 
-        for (Integer idSPCT : listIDSPCT) {
-            SanPhamChiTiet sanPhamChiTiet = sanPhamChiTietRepo.getReferenceById(idSPCT);
+            List<GioHangChiTiet> gioHangChiTiets = gioHangChiTietRepo.gioHangChiTietByID(idGioHang);
+            List<Integer> listIDSPCT = gioHangChiTietRepo.findIdSanPhamChiTietByIdGioHang(idGioHang);
 
-            HoaDonChiTiet hoaDonChiTietOnl = new HoaDonChiTiet();
-            hoaDonChiTietOnl.setHoaDon(hoaDon);
-            hoaDonChiTietOnl.setNgayTao(new Date());
-            hoaDonChiTietOnl.setNgaySua(new Date());
-            hoaDonChiTietOnl.setTrangThai(1);
-            hoaDonChiTietOnl.setSanPhamChiTiet(sanPhamChiTiet);
+            for (Integer idSPCT : listIDSPCT) {
+                SanPhamChiTiet sanPhamChiTiet = sanPhamChiTietRepo.getReferenceById(idSPCT);
 
-            for (GioHangChiTiet gioHangChiTiet : gioHangChiTiets) {
-                hoaDonChiTietOnl.setSoLuong(gioHangChiTiet.getSoLuong());
+                HoaDonChiTiet hoaDonChiTietOnl = new HoaDonChiTiet();
+                hoaDonChiTietOnl.setHoaDon(hoaDon);
+                hoaDonChiTietOnl.setNgayTao(new Date());
+                hoaDonChiTietOnl.setNgaySua(new Date());
+                hoaDonChiTietOnl.setTrangThai(1);
+                hoaDonChiTietOnl.setSanPhamChiTiet(sanPhamChiTiet);
+
+                for (GioHangChiTiet gioHangChiTiet : gioHangChiTiets) {
+                    hoaDonChiTietOnl.setSoLuong(gioHangChiTiet.getSoLuong());
+                }
+
+                hoaDonChiTietRepo.save(hoaDonChiTietOnl);
+
             }
 
-            hoaDonChiTietRepo.save(hoaDonChiTietOnl);
+
+            session.setAttribute("giaTriGiam", 0);
+
+            gioHangChiTietRepo.deleteByIdGioHang(idGioHang);
+        }
+        return "redirect:/t-shirt-luxury/trang-chu";
+    }
+
+    @PostMapping("/t-shirt-luxury/ban-hang-onl/update-quantity")
+    @ResponseBody // Trả về JSON khi được gọi bằng AJAX
+    public ResponseEntity<String> updateQuantity(@RequestParam("idGioHangChiTiet") Integer idGioHangChiTiet,
+                                                 @RequestParam("soLuong") Integer soLuong) {
+        try {
+            GioHangChiTiet gioHangChiTiet = gioHangChiTietRepo.findById(idGioHangChiTiet)
+                    .orElseThrow(() -> new IllegalArgumentException("Chi tiết giỏ hàng không tồn tại"));
+
+            gioHangChiTiet.setSoLuong(soLuong);
+            gioHangChiTietRepo.save(gioHangChiTiet);
+            System.out.println(idGioHangChiTiet);
+
+            return ResponseEntity.ok("Cập nhật thành công!");
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("Cập nhật thất bại!"+idGioHangChiTiet);
 
         }
-
-
-        session.setAttribute("giaTriGiam", 0);
-
-        gioHangChiTietRepo.deleteByIdGioHang(idGioHang);
-
-        return "redirect:/t-shirt-luxury/trang-chu";
     }
